+++ fastly/ngwaf/v1/workspaces/lists/api_list.go
@@ -1,0 +1,40 @@
+package lists
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+
+	"github.com/fastly/go-fastly/v10/fastly"
+)
+
+// ListInput specifies the information needed for the List() function to perform
+// the operation.
+type ListInput struct {
+	// Context, if supplied, will be used as the Request's context.
+	Context *context.Context
+	// WorkspaceID is the workspace identifier (required).
+	WorkspaceID *string
+}
+
+// ListLists retrieves a list of lists for the given workspace.
+func ListLists(c *fastly.Client, i *ListInput) (*Lists, error) {
+	if i.WorkspaceID == nil {
+		return nil, fastly.ErrMissingWorkspaceID
+	}
+
+	path := fastly.ToSafeURL("ngwaf", "v1", "workspaces", *i.WorkspaceID, "lists")
+
+	resp, err := c.Get(path, fastly.CreateRequestOptions(i.Context))
+	if err != nil {
+		return nil, err
+	}
+	defer resp.Body.Close()
+
+	var lists *Lists
+	if err := json.NewDecoder(resp.Body).Decode(&lists); err != nil {
+		return nil, fmt.Errorf("failed to decode json response: %w", err)
+	}
+
+	return lists, nil
+}